requirements.txt
fastapi==0.115.6
httpx==0.27.2
python-dotenv==1.0.1
runpod==1.4.2
uvicorn[standard]==0.34.0
websockets==12.0
